+++ packages/dart/lib/src/scope_observer.dart
@@ -6,9 +6,9 @@ import 'protocol/sentry_user.dart';
 abstract class ScopeObserver {
   Future<void> setContexts(String key, dynamic value);
   Future<void> removeContexts(String key);
-  Future<void> setUser(SentryUser? user);
-  Future<void> addBreadcrumb(Breadcrumb breadcrumb);
-  Future<void> clearBreadcrumbs();
+  FutureOr<void> setUser(SentryUser? user);
+  FutureOr<void> addBreadcrumb(Breadcrumb breadcrumb);
+  FutureOr<void> clearBreadcrumbs();
   Future<void> setExtra(String key, dynamic value);
   Future<void> removeExtra(String key);
   Future<void> setTag(String key, String value);