+++ packages/flutter/lib/src/native/cocoa/sentry_native_cocoa.dart
@@ -191,4 +191,73 @@ class SentryNativeCocoa extends SentryNativeChannel {
           return NativeAppStart.fromJson(json);
         },
       );
+
+  @override
+  void addBreadcrumb(Breadcrumb breadcrumb) =>
+      tryCatchSync('addBreadcrumb', () {
+        final nativeBreadcrumb =
+            cocoa.PrivateSentrySDKOnly.breadcrumbWithDictionary(
+                _deepConvertMapNonNull(breadcrumb.toJson()).toNSDictionary());
+        cocoa.SentrySDK.addBreadcrumb(nativeBreadcrumb);
+      });
+
+  @override
+  void clearBreadcrumbs() => tryCatchSync('clearBreadcrumbs', () {
+        cocoa.SentrySDK.configureScope(
+            cocoa.ObjCBlock_ffiVoid_SentryScope.fromFunction(
+                (cocoa.SentryScope scope) {
+          scope.clearBreadcrumbs();
+        }));
+      });
+
+  @override
+  void nativeCrash() => cocoa.SentrySDK.crash();
+
+  @override
+  void pauseAppHangTracking() => tryCatchSync('pauseAppHangTracking', () {
+        cocoa.SentrySDK.pauseAppHangTracking();
+      });
+
+  @override
+  void resumeAppHangTracking() => tryCatchSync('resumeAppHangTracking', () {
+        cocoa.SentrySDK.resumeAppHangTracking();
+      });
+
+  @override
+  void setUser(SentryUser? user) => tryCatchSync('setUser', () {
+        if (user == null) {
+          cocoa.SentrySDK.setUser(null);
+        } else {
+          final dictionary =
+              _deepConvertMapNonNull(user.toJson()).toNSDictionary();
+          final cUser =
+              cocoa.PrivateSentrySDKOnly.userWithDictionary(dictionary);
+          cocoa.SentrySDK.setUser(cUser);
+        }
+      });
+}
+
+/// This map conversion is needed so we can use the toNSDictionary extension function
+/// provided by the objective_c package.
+Map<Object, Object> _deepConvertMapNonNull(Map<String, dynamic> input) {
+  final out = <Object, Object>{};
+
+  for (final entry in input.entries) {
+    final value = entry.value;
+    if (value == null) continue;
+
+    out[entry.key] = switch (value) {
+      Map<String, dynamic> m => _deepConvertMapNonNull(m),
+      List<dynamic> l => [
+          for (final e in l)
+            if (e != null)
+              e is Map<String, dynamic>
+                  ? _deepConvertMapNonNull(e)
+                  : e as Object
+        ],
+      _ => value as Object,
+    };
+  }
+
+  return out;
 }